+++ example/main.dart
@@ -1,12 +1,17 @@
 import 'package:flutter/material.dart';
-import 'package:linear_gradient/src/linear_gradient_style.dart';
+
+import '../lib/linear_gradient.dart';
 
 void main() => runApp(MyApp());
 
-class MyApp extends StatelessWidget {
+class MyApp extends StatefulWidget {
+  @override
+  _MyAppState createState() => _MyAppState();
+}
+
+class _MyAppState extends State<MyApp> {
   /// This is an example app which make use of
   /// How to use `Custom Gradient Box Decoration`
-
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
@@ -18,4 +23,8 @@ class MyApp extends StatelessWidget {
       ),
     );
   }
+
+  /// Get Gradients color list by passing gradient type
+  List<Color> colorCombination = LinearGradientStyle.getColorCombination(
+      gradientType: LinearGradientStyle.GRADIENT_TYPE_ROYAL);
 }