+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 0.1.10
+
+- Added Documents.
+- Update README file.
+
+## 0.1.9
+
+- Added new method getColorCombination() to get Gradients color list by passing gradient type.
+- Now you can use Gradients style color as theme. 
+
+## 0.1.8
+
+- Update README.
+
 ## 0.1.7
 
 - Added Screenshot in README.